chore: clear the remaining knip warnings
Three no-behaviour cleanups:

- currencies.ts exported the list both named and default, but only the
  default is consumed (index.ts lazy-imports it). Drop the named export.
- src/logger/index.ts and src/try/index.ts were one-line re-export barrels;
  the implementation files now are the index. Same public API, two fewer
  files. Verified ./try and ./logger still resolve from a real build.
- knip.json's entry and project patterns were the defaults, and knip itself
  reported all three as redundant.

Closes #221
