+++ knip.json
@@ -1,5 +1,3 @@
 {
-  "$schema": "https://unpkg.com/knip@5/schema.json",
-  "entry": ["src/index.ts"],
-  "project": ["src/**/*.ts"]
+  "$schema": "https://unpkg.com/knip@5/schema.json"
 }
